backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
email-validator
gunicorn

